# apps/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Completionist application.",
        "title": "Completionist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/challenges": {
            "get": {
                "description": "Retrieves list of challenges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Challenges"
                ],
                "summary": "Get Challenges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by frequency (daily, weekly, etc)",
                        "name": "frequency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Challenge"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/lists/google/books/{book_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports a book from Google Books into the user's list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lists"
                ],
                "summary": "Add Book from Google Books",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Google Books Volume ID",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addFromGoogleBooksBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserListItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/me/challenges": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the current user's progress on challenges",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Challenges"
                ],
                "summary": "Get My Challenge Progress",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserChallenge"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/search/books": {
            "get": {
                "description": "Search for books using the Google Books API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search Books (Google Books)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/googlebooks.SearchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/seasons/current": {
            "get": {
                "description": "Retrieves the current active season information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Get Active Season",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Season"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seasons/leaderboard": {
            "get": {
                "description": "Retrieves the leaderboard for the current season based on XP gained",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Get Season Leaderboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.LeaderboardEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "googlebooks.Book": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "volumeInfo": {
                    "$ref": "#/definitions/googlebooks.VolumeInfo"
                }
            }
        },
        "googlebooks.ImageLinks": {
            "type": "object",
            "properties": {
                "smallThumbnail": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "googlebooks.SearchResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/googlebooks.Book"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "totalItems": {
                    "type": "integer"
                }
            }
        },
        "googlebooks.VolumeInfo": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "imageLinks": {
                    "$ref": "#/definitions/googlebooks.ImageLinks"
                },
                "pageCount": {
                    "type": "integer"
                },
                "publishedDate": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.addFromGoogleBooksBody": {
            "type": "object",
            "properties": {
                "progressData": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ItemStatus"
                }
            }
        },
...
        "models.Challenge": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "criteriaMetadata": {
                    "type": "string"
                },
                "criteriaType": {
                    "$ref": "#/definitions/models.CriteriaType"
                },
                "description": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/models.ChallengeFrequency"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.ChallengeType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "xpReward": {
                    "type": "integer"
                }
            }
        },
        "models.ChallengeFrequency": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "monthly",
                "seasonal",
                "infinite"
            ],
            "x-enum-varnames": [
                "FrequencyDaily",
                "FrequencyWeekly",
                "FrequencyMonthly",
                "FrequencySeasonal",
                "FrequencyInfinite"
            ]
        },
        "models.ChallengeType": {
            "type": "string",
            "enum": [
                "global",
                "personal"
            ],
            "x-enum-varnames": [
                "ChallengeTypeGlobal",
                "ChallengeTypePersonal"
            ]
        },
...
        "models.CriteriaType": {
            "type": "string",
            "enum": [
                "count_items",
                "specific_item",
                "genre_count"
            ],
            "x-enum-varnames": [
                "CriteriaCountItems",
                "CriteriaSpecificItem",
                "CriteriaGenreCount"
            ]
        },
...
        "models.ItemStatus": {
            "type": "string",
            "enum": [
                "planning",
                "current",
                "completed",
                "paused",
                "dropped"
            ],
            "x-enum-varnames": [
                "StatusPlanning",
                "StatusCurrent",
                "StatusCompleted",
                "StatusPaused",
                "StatusDropped"
            ]
        },
        "models.ItemType": {
            "type": "string",
            "enum": [
                "manga",
                "game",
                "book",
                "movie",
                "series",
                "music"
            ],
            "x-enum-varnames": [
                "ItemTypeManga",
                "ItemTypeGame",
                "ItemTypeBook",
                "ItemTypeMovie",
                "ItemTypeSeries",
                "ItemTypeMusic"
            ]
        },
...
        "models.MediaItem": {
            "type": "object",
            "properties": {
                "coverImageUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "criticRatingCount": {
                    "type": "integer"
                },
                "criticRatingValue": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "externalId": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "itemType": {
                    "$ref": "#/definitions/models.ItemType"
                },
                "metadata": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userRatingExternal": {
                    "type": "number"
                }
            }
        },
...
        "models.Season": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "endAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "startAt": {
                    "type": "string"
                }
            }
        },
...
        "models.UserChallenge": {
            "type": "object",
            "properties": {
                "challengeDescription": {
                    "type": "string"
                },
                "challengeId": {
                    "type": "integer"
                },
                "challengeTitle": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentProgress": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "seasonId": {
                    "type": "integer"
                },
                "targetProgress": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "xpReward": {
                    "type": "integer"
                }
            }
        },
...
        "models.UserListItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itemType": {
                    "$ref": "#/definitions/models.ItemType"
                },
                "media": {
                    "$ref": "#/definitions/models.MediaItem"
                },
                "mediaItemId": {
                    "type": "string"
                },
                "progressData": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.ItemStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "repository.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "rank": {
                    "type": "integer"
                },
                "totalXp": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}